create PSPageBlobClient.cs in alcWrapper and move content to it
