+++ src/Compute/Compute.AlcWrapper/AzureStorageService.cs
@@ -14,7 +14,6 @@
 
 using Azure.Storage.Blobs;
 using Azure.Storage.Blobs.Models;
-using Azure.Storage.Blobs.Specialized;
 using System.IO;
 using System;
 using System.Threading;
@@ -59,26 +58,5 @@ namespace Microsoft.Azure.Commands.Compute
         public Stream Content { get { return _blobDownloadInfo.Content; } }
 
         public string ContentType { get { return _blobDownloadInfo.ContentType; } }
-    }
-
-    public class PSPageBlobClient
-    {
-        private PageBlobClient _pageBlobClient;
-
-        internal PSPageBlobClient(PageBlobClient pageblobclient)
-        {
-            _pageBlobClient = pageblobclient;
-        }
-        public PSPageBlobClient(Uri blobUri)
-        {
-            _pageBlobClient = new PageBlobClient(blobUri, null);
-        }
-
-        public Uri Uri { get { return _pageBlobClient.Uri; } }
-
-        public void UploadPages(Stream content, long offset)
-        {
-            _pageBlobClient.UploadPagesAsync(content, offset).ConfigureAwait(false).GetAwaiter().GetResult();
-        }
-    }
+    } 
 }
